This is the first commit for this learning project.
I setup a local Postgres Server and created database tables outside of this project and commit.

I created the flask project with the following structure:

   app/
       templates/
             insert_weight.html
       __init__.py
       forms.py
       routes.py
  .gitignore
  config.py
  Weight.py

Forms.py contains the AddWeightForm that has a StringField for every day of the week. The form validates that at least, one weight was
entered for the week, and that the value is a valid number or at least, can be parsed into a number by Python float().

The AddWeightForm is served on '/' and in the function index() and since this is currently the only thing I have developed for my project.
Also, index calculates the dates of the week with the first day of the week being Sunday. It calls on the template to display by passing in the form
and the dates. Once the form is submit, it calls the same location '/' with a post method and validates the input. If at least one day has valid data,
it is loaded into the database.

I'm using flask_sqlalchemy, but haven't quite figured out the models portion yet. I work as Report Developer and thus, I'm very comfortable with SQL.
Eventually, I'll make use of more of the ORM capabilities.