I updated my AddWeightForm to use a FieldList of StringFields instead of creating a StringField for each individual day of the week. I updated the validation in the AddWeightForm to reflect this and the insert_weight.html template.
In routes.py I updated the index() function due to the change to using FieldList for the days and added the ability to update or delete weights for the week.
I also updated the sql query to have postgres generate the days with no values instead of filling it out in Python.

I also changed the data table I was using and added the SQL I used to generate the new data source.